Delegate Bento methods to config accessor instead of ivar
It's possible there's a reason we don't want these methods to install a default configuration by side-effect. But barring that, the delegation errors we get when calling these methods in the absence of a configuration are confusing and not super helpful.

If we want to avoid setting @config implicitly, I can pitch a different PR